Isolate mask compositing from engine facade
apply_mask is a focused alphamerge transform and can move into a dedicated module while preserving mcp_video.engine.apply_mask for public callers. The extracted module reuses shared quality args and escapes numeric filter values before filter_complex interpolation.

Constraint: Keep existing mask behavior and audio-preservation semantics while continuing engine decomposition.

Rejected: Combine with overlay/split-screen extraction | those paths have separate composition semantics and should be reviewed in their own slices.

Confidence: high

Scope-risk: narrow

Directive: Mask filter_complex inputs must stay sanitized before interpolation.

Tested: ruff check mcp_video/engine.py mcp_video/engine_mask.py; ruff format --check mcp_video/engine.py mcp_video/engine_mask.py; python3 -m pytest tests/test_engine_advanced.py -k 'apply_mask' -q --tb=short; python3 -m pytest tests/test_cli.py -k 'apply_mask' tests/test_server.py::TestVideoApplyMaskTool tests/test_red_team.py -k 'apply_mask' -q --tb=short; python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short

Not-tested: Full slow/real-media suite.
